rho falls to zero, why?

+++ cavity_params.py
@@ -25,32 +25,32 @@ reynolds_lattice = reynolds     # keep same reynolds
 viscosity_lattice = 0.01        # choose so we keep same reynolds
 velocity_lattice = 0.1          # choose so we keep same reynolds (here we ignore the actual velocity)
 
-omega = 1./(3. * viscosity_lattice + 0.5)    # TODO is omega correct here? I get omega = 3*v_lattice + 1/2
+omega = 1./(3. * viscosity_lattice + 0.5)
 #omega = (3. * velocity_lattice) + 0.5
 
 N = int(round(reynolds_lattice * viscosity_lattice / velocity_lattice))     # lattices in x-direction
 M = int(round(N / aspect))                                                  # lattices in y-direction (want dx = dy)
 
 # weights for D2Q9
-w = np.zeros(9,dtype=float)
+w = np.zeros(9)
 w[0] = 4./9.
 w[1:5] = 1./9.
 w[5:9] = 1./36.
 
 # grids MxNx9
-f = np.zeros((9,M,N),dtype=float)       # distribution function
-rho = np.zeros((M,N),dtype=float)       # density
-u = np.zeros((2,M,N),dtype=float)       # velocity
+f = np.zeros((9,M,N))       # distribution function
+rho = np.zeros((M,N))       # density
+u = np.zeros((2,M,N))       # velocity
 
 # boundary condition
-u0 = np.array((velocity_lattice,0.),dtype=float)
+u0 = np.array((velocity_lattice,0.))
 u[0,0,1:-1] = velocity_lattice     # load BC into north boundary
 
 # initial condition
 rho[:,:] = 5.
 
 # unit vectors for D2Q9
-c = np.array(((0,0),(1,0),(0,1),(-1,0),(0,-1),(1,1),(-1,1),(-1,-1),(1,-1)),dtype=float)
+c = np.array(((0,0),(1,0),(0,1),(-1,0),(0,-1),(1,1),(-1,1),(-1,-1),(1,-1)))
